difan-backend/difan_backend/requirements.txt
dagster
dagster-postgres
dagster-docker

mysql-connector-python
requests
pandas
numpy

yfinance
bs4
beautifulsoup4

torch
regex
toml
transformers
fuzzywuzzy
python-levenshtein